fix: Merge conflict and changes
